+++ packages/tools/src/setup-registry.ts
@@ -1,35 +1,35 @@
 import { Registry } from "@radon-lite/registry";
 import { simulatorServerBlueprint } from "./blueprints/simulator-server";
 import { metroDebuggerBlueprint } from "./blueprints/metro-debugger";
-import { listSimulatorsTool } from "./tools/list-simulators";
-import { bootSimulatorTool } from "./tools/boot-simulator";
-import { simulatorServerTool } from "./tools/simulator-server";
-import { launchAppTool } from "./tools/launch-app";
-import { openUrlTool } from "./tools/open-url";
-import { screenshotTool } from "./tools/screenshot";
-import { tapTool } from "./tools/tap";
-import { swipeTool } from "./tools/swipe";
-import { gestureTool } from "./tools/gesture";
-import { buttonTool } from "./tools/button";
-import { keyboardTool } from "./tools/keyboard";
-import { rotateTool } from "./tools/rotate";
-import { metroConnectTool } from "./tools/metro-connect";
-import { metroStatusTool } from "./tools/metro-status";
-import { metroEvaluateTool } from "./tools/metro-evaluate";
-import { metroSetBreakpointTool } from "./tools/metro-set-breakpoint";
-import { metroRemoveBreakpointTool } from "./tools/metro-remove-breakpoint";
-import { metroPauseTool } from "./tools/metro-pause";
-import { metroResumeTool } from "./tools/metro-resume";
-import { metroStepTool } from "./tools/metro-step";
-import { metroComponentTreeTool } from "./tools/metro-component-tree";
-import { metroInspectElementTool } from "./tools/metro-inspect-element";
-import { metroConsoleLogsTool } from "./tools/metro-console-logs";
-import { metroConsoleListenTool } from "./tools/metro-console-listen";
-import { describeTool } from "./tools/describe";
-import { activateLicenseKeyTool } from "./tools/activate-license-key";
-import { activateSsoTool } from "./tools/activate-sso";
-import { getLicenseStatusTool } from "./tools/get-license-status";
-import { removeLicenseTool } from "./tools/remove-license";
+import { listSimulatorsTool } from "./tools/simulator/list-simulators";
+import { bootSimulatorTool } from "./tools/simulator/boot-simulator";
+import { simulatorServerTool } from "./tools/simulator/simulator-server";
+import { launchAppTool } from "./tools/simulator/launch-app";
+import { openUrlTool } from "./tools/simulator/open-url";
+import { screenshotTool } from "./tools/interactions/screenshot";
+import { tapTool } from "./tools/interactions/tap";
+import { swipeTool } from "./tools/interactions/swipe";
+import { gestureTool } from "./tools/interactions/gesture";
+import { buttonTool } from "./tools/interactions/button";
+import { keyboardTool } from "./tools/interactions/keyboard";
+import { rotateTool } from "./tools/simulator/rotate";
+import { metroConnectTool } from "./tools/metro/metro-connect";
+import { metroStatusTool } from "./tools/metro/metro-status";
+import { metroEvaluateTool } from "./tools/metro/metro-evaluate";
+import { metroSetBreakpointTool } from "./tools/metro/metro-set-breakpoint";
+import { metroRemoveBreakpointTool } from "./tools/metro/metro-remove-breakpoint";
+import { metroPauseTool } from "./tools/metro/metro-pause";
+import { metroResumeTool } from "./tools/metro/metro-resume";
+import { metroStepTool } from "./tools/metro/metro-step";
+import { metroComponentTreeTool } from "./tools/metro/metro-component-tree";
+import { metroInspectElementTool } from "./tools/metro/metro-inspect-element";
+import { metroConsoleLogsTool } from "./tools/metro/metro-console-logs";
+import { metroConsoleListenTool } from "./tools/metro/metro-console-listen";
+import { describeTool } from "./tools/interactions/describe";
+import { activateLicenseKeyTool } from "./tools/license/activate-license-key";
+import { activateSsoTool } from "./tools/license/activate-sso";
+import { getLicenseStatusTool } from "./tools/license/get-license-status";
+import { removeLicenseTool } from "./tools/license/remove-license";
 
 export function createRegistry(): Registry {
   const registry = new Registry();
