+++ packages/tools/src/tools/interactions/keyboard.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import type { ToolDefinition } from "@radon-lite/registry";
-import type { SimulatorServerApi } from "../blueprints/simulator-server";
+import type { SimulatorServerApi } from "../../blueprints/simulator-server";
 
 const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
