+++ packages/tools/src/tools/metro/metro-console-logs.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import type { ToolDefinition } from "@radon-lite/registry";
-import type { MetroDebuggerApi, ConsoleLogEntry } from "../blueprints/metro-debugger";
+import type { MetroDebuggerApi, ConsoleLogEntry } from "../../blueprints/metro-debugger";
 
 const zodSchema = z.object({
   port: z.number().default(8081).describe("Metro server port"),
