+++ packages/tools/src/tools/metro/metro-status.ts
@@ -1,12 +1,12 @@
 import { z } from "zod";
 import type { ToolDefinition } from "@radon-lite/registry";
-import type { MetroDebuggerApi } from "../blueprints/metro-debugger";
+import type { MetroDebuggerApi } from "../../blueprints/metro-debugger";
 
 const zodSchema = z.object({
   port: z.number().default(8081).describe("Metro server port"),
 });
 
-export const metroConnectTool: ToolDefinition<
+export const metroStatusTool: ToolDefinition<
   z.infer<typeof zodSchema>,
   {
     port: number;
@@ -14,24 +14,31 @@ export const metroConnectTool: ToolDefinition<
     deviceName: string;
     isNewDebugger: boolean;
     connected: boolean;
+    loadedScripts: number;
+    enabledDomains: string[];
+    sourceMapReady: boolean;
   }
 > = {
-  id: "metro-connect",
-  description: `Connect to a running Metro dev server's CDP debugger endpoint.
-Returns connection info. If already connected, returns the existing connection.
-This must be called before using other metro-* tools (or they auto-connect).`,
+  id: "metro-status",
+  description: `Get Metro debugger connection status and diagnostic info.
+Connects to Metro if not already connected (idempotent with metro-connect).
+Includes source map readiness status for breakpoint resolution.`,
   zodSchema,
   services: (params) => ({
     metroDebugger: `MetroDebugger:${params.port}`,
   }),
   async execute(services) {
     const api = services.metroDebugger as MetroDebuggerApi;
+    await api.sourceMaps.waitForPending();
     return {
       port: api.port,
       projectRoot: api.projectRoot,
       deviceName: api.deviceName,
       isNewDebugger: api.isNewDebugger,
       connected: api.cdp.isConnected(),
+      loadedScripts: api.cdp.getLoadedScripts().size,
+      enabledDomains: [...api.cdp.getEnabledDomains()],
+      sourceMapReady: true,
     };
   },
 };
