+++ packages/tools/src/tools/simulator/boot-simulator.ts
