+++ packages/tools/src/tools/simulator/simulator-server.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import type { ToolDefinition } from "@radon-lite/registry";
-import type { SimulatorServerApi } from "../blueprints/simulator-server";
+import type { SimulatorServerApi } from "../../blueprints/simulator-server";
 
 const zodSchema = z.object({
   udid: z.string().describe("The UDID of the simulator to connect to"),
